refactor: reorganize state machine files and update example usage
