+++ core/parser_machine.py
@@ -203,38 +203,3 @@ class StateMachineParser:
         except Exception as e:
             logger.error(f"Error: {str(e)}")
             raise
-
-
-def use_example_parallel():
-    """Run the example parallel state machine with a sample event."""
-
-    parser_handler = StateMachineParser('sm_p_description.yml')
-    machine = parser_handler.parse()
-
-    if not machine:
-        return
-
-    event: dict[str, int] = {"value": 50}
-    result = machine.run(event)
-    print(result)
-
-
-def use_example():
-    """Run the example serial state machine with a sample event."""
-
-    parser_handler = StateMachineParser('sm_description.yml')
-    machine = parser_handler.parse()
-
-    if not machine:
-        return
-
-    event: dict[str, int] = {"value": 50}
-    result = machine.run(event)
-    print("RESULT: ", result)
-
-
-if __name__ == "__main__":
-    print("Serial...")
-    use_example()
-    print("\nParallel...")
-    use_example_parallel()
