+++ machines/sm_description.yml
