+++ machines/sm_p_description.yml
